hide helper traits from calling code
Makes a couple of changes to avoid helper traits from appearing anywhere in the
caller code.

It changes the name of the `Display` trait method to reduce the chance that an
inherent method conflicts with the type in question, causing undesirable
results. It also wraps the helper trait and declarations into an anonymous const
context, meaning the helper traits aren't visible to the calling context.